Improve __validate__ performance  (#164)

* read input CSV in batches of 16k, and process each batch using Rayon parallel iterator
* collect validation results to write to validation error file after main loop
* split CSV into valid/invalid files at the end based on Vec<bool>. assumes Rayon maintains order of original CSV file
* using Rayon instead of threadpool, progress bar can still be accurately updated.

I got about 2X improvement on my machine.

__single thread__
```
(base) ➜  qsv git:(validate_performance) ✗ time target/release/qsvlite validate tmp/NYC_311_SR_2010-2020-sample-1M.csv tmp/NYC-short.csv.schema.json
[00:01:20] [==================== 100% validated 1,000,000 records.] (12,393/sec)
6,494 out of 1,000,000 records invalid.
target/release/qsvlite validate tmp/NYC_311_SR_2010-2020-sample-1M.csv   79.87s user 0.70s system 99% cpu 1:20.87 total
```

__Rayon parallel iterator__
```
(base) ➜  qsv git:(validate_performance) ✗ time ./target/release/qsvlite validate tmp/NYC_311_SR_2010-2020-sample-1M.csv tmp/NYC-short.csv.schema.json
[00:00:41] [==================== 100% validated 1,000,000 records.] (28,850/sec)
6,494 out of 1,000,000 records invalid.
./target/release/qsvlite validate tmp/NYC_311_SR_2010-2020-sample-1M.csv   130.22s user 0.70s system 317% cpu 41.252 total
```