@jqnatividad 2022-02-19
Awesome @mhuang74! 
progress bar being enabled is a bonus too (and that's before taking advantage of [`indicatif`'s rayon support](https://docs.rs/indicatif/0.16.2/indicatif/index.html#iterators))!  This will be the reference implementation for moving the current parallelized commands to rayon, but also, parallelizing more commands going forward.